Shared LMS cookie should be shared across all pages
Fixes an issue where the LMS would return an error 500 in certain conditions.

The issue was that the shared session cookie path was set to the first page from the client UI on which an XBlock was loaded, instead of `/`, which would prevent the session id to be passed, and resulting in not being logged in on the LMS, which would crash certain XBlocks.